add nemea-siot packages
SIoT is a project in cooperation of CESNET and Technical University of
Ostrava.

The goal of this project is to build a secured gateway for IoT devices.
The gateway consists of various monitoring and detection
functionalities.

There are some detection modules developed in this project and located
in https://github.com/CESNET/NEMEA-SIoT

This PR adds the modules into build system for OpenWrt.