travis: Remove it entirely
I propose to remove Travis, because it tries to compile and install
packages for Turris OS 3.x [1]. This version is in maintenance mode, so it
gets security updates and receives small various bug fixes. We would like to push our users
a little bit to try opt-in migration [2] to Turris OS 5.x, which is based on the latest OpenWrt
stable release until we do it automatically.

If you are using Turris 1.x routers, you will need to have microSD card
inserted in your router to jump on Turris OS 5.x and with Btrfs, you
will enjoy snapshots.

If you are using Turris Omnia, there is a possibility to doing re-flash,
but opt-in migration is that what we want.

Also, users on Turris OS 5.x will benefit from the newer version of Python3
among other things.

[1] https://docs.turris.cz/basics/tos_versions/
[2] https://docs.turris.cz/geek/tos3_migration/#optional-migration